requirements.txt
torch
torchvision
transformers
datasets
accelerate
scikit-learn
pillow
timm
